+++ packages/playground/src/index.js
@@ -1,3 +1,7 @@
+import 'core-js/features/array/fill';
+import 'core-js/features/math/sign';
+import 'react-app-polyfill/ie11';
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';